requirements.txt
pandas==0.23.4
pyodbc==4.0.24
python_dateutil==2.7.3
python-pptx==0.6.6